requirements.txt
Flask
flask-cors
joblib
SpeechRecognition
Werkzeug
openai-whisper
pytest
